hotfix: disable Privy SSR — dynamic import with ssr:false
CRITICAL: Privy SDK crashes during server-side rendering because it
accesses browser-only APIs (window, localStorage). The error boundary
from PR #326 only catches client-side errors, not SSR crashes.

Fix: Split PrivyProvider into a separate PrivyProviderClient component
and load it via next/dynamic with { ssr: false }. During SSR, children
render without wallet capability. Privy loads client-side after hydration.

- New: PrivyProviderClient.tsx (client-only Privy initialization)
- WalletProvider.tsx: uses dynamic() import instead of direct import
