+++ app/components/providers/PrivyProviderClient.tsx
@@ -1,0 +1,66 @@
+"use client";
+
+import { FC, ReactNode, useMemo } from "react";
+import { PrivyProvider, usePrivy } from "@privy-io/react-auth";
+import { toSolanaWalletConnectors } from "@privy-io/react-auth/solana";
+import { SentryUserContext } from "@/components/providers/SentryUserContext";
+import { PrivyLoginContext } from "@/hooks/usePrivySafe";
+import { getConfig } from "@/lib/config";
+
+/**
+ * Client-only Privy provider wrapper. Loaded via next/dynamic with ssr:false
+ * to prevent Privy SDK from crashing during server-side rendering.
+ */
+const PrivyProviderClient: FC<{ appId: string; children: ReactNode }> = ({
+  appId,
+  children,
+}) => {
+  const rpcUrl = useMemo(() => {
+    const url = getConfig().rpcUrl;
+    if (!url || !url.startsWith("http")) return "https://api.devnet.solana.com";
+    return url;
+  }, []);
+
+  const solanaConnectors = useMemo(() => toSolanaWalletConnectors(), []);
+
+  return (
+    <PrivyProvider
+      appId={appId}
+      config={{
+        appearance: {
+          walletChainType: "solana-only",
+          showWalletLoginFirst: true,
+        },
+        loginMethods: ["wallet", "email"],
+        externalWallets: {
+          solana: {
+            connectors: solanaConnectors,
+          },
+        },
+        embeddedWallets: {
+          solana: {
+            createOnLogin: "users-without-wallets",
+          },
+        },
+      }}
+    >
+      <SentryUserContext />
+      <PrivyLoginBridge>{children}</PrivyLoginBridge>
+    </PrivyProvider>
+  );
+};
+
+/**
+ * Bridge that exposes Privy's login function via context so components
+ * outside the Privy tree can trigger wallet connection safely.
+ */
+const PrivyLoginBridge: FC<{ children: ReactNode }> = ({ children }) => {
+  const { login } = usePrivy();
+  return (
+    <PrivyLoginContext.Provider value={login}>
+      {children}
+    </PrivyLoginContext.Provider>
+  );
+};
+
+export default PrivyProviderClient;
