+++ app/components/providers/WalletProvider.tsx
@@ -1,11 +1,8 @@
 "use client";
 
-import { Component, FC, ReactNode, useMemo } from "react";
-import { PrivyProvider, usePrivy } from "@privy-io/react-auth";
-import { toSolanaWalletConnectors } from "@privy-io/react-auth/solana";
-import { SentryUserContext } from "@/components/providers/SentryUserContext";
+import { Component, FC, ReactNode, useMemo, useState, useEffect } from "react";
+import dynamic from "next/dynamic";
 import { PrivyAvailableContext, PrivyLoginContext } from "@/hooks/usePrivySafe";
-import { getConfig } from "@/lib/config";
 
 /**
  * Error boundary that catches PrivyProvider crashes and renders children
@@ -40,19 +37,18 @@ class PrivyErrorBoundary extends Component<
   }
 }
 
+/**
+ * Dynamically import the Privy wrapper with SSR disabled.
+ * This prevents Privy SDK from being evaluated during server-side rendering,
+ * which crashes because Privy accesses browser-only APIs (window, localStorage).
+ */
+const PrivyProviderClient = dynamic(
+  () => import("./PrivyProviderClient").then((mod) => mod.default),
+  { ssr: false }
+);
+
 export const WalletProvider: FC<{ children: ReactNode }> = ({ children }) => {
   const appId = process.env.NEXT_PUBLIC_PRIVY_APP_ID;
-
-  // If no Privy app ID is configured, skip PrivyProvider entirely.
-  // The app runs in read-only mode (no wallet connection).
-  // PrivyAvailableContext = false tells all hooks to return defaults.
-  if (!appId) {
-    return (
-      <PrivyAvailableContext.Provider value={false}>
-        {children}
-      </PrivyAvailableContext.Provider>
-    );
-  }
 
   const readOnlyFallback = (
     <PrivyAvailableContext.Provider value={false}>
@@ -60,69 +56,19 @@ export const WalletProvider: FC<{ children: ReactNode }> = ({ children }) => {
     </PrivyAvailableContext.Provider>
   );
 
+  // No app ID configured: read-only mode (skip Privy entirely)
+  if (!appId) {
+    return readOnlyFallback;
+  }
+
+  // Mount Privy client-side only via dynamic import (ssr: false)
   return (
     <PrivyErrorBoundary fallback={readOnlyFallback}>
       <PrivyAvailableContext.Provider value={true}>
-        <PrivyProviderWrapper appId={appId}>
+        <PrivyProviderClient appId={appId}>
           {children}
-        </PrivyProviderWrapper>
+        </PrivyProviderClient>
       </PrivyAvailableContext.Provider>
     </PrivyErrorBoundary>
-  );
-};
-
-/**
- * Inner component that initializes Privy. Separated so the error boundary
- * can catch any errors thrown during PrivyProvider initialization or rendering.
- */
-const PrivyProviderWrapper: FC<{ appId: string; children: ReactNode }> = ({
-  appId,
-  children,
-}) => {
-  const rpcUrl = useMemo(() => {
-    const url = getConfig().rpcUrl;
-    if (!url || !url.startsWith("http")) return "https://api.devnet.solana.com";
-    return url;
-  }, []);
-
-  const solanaConnectors = useMemo(() => toSolanaWalletConnectors(), []);
-
-  return (
-    <PrivyProvider
-      appId={appId as string}
-      config={{
-        appearance: {
-          walletChainType: "solana-only",
-          showWalletLoginFirst: true,
-        },
-        loginMethods: ["wallet", "email"],
-        externalWallets: {
-          solana: {
-            connectors: solanaConnectors,
-          },
-        },
-        embeddedWallets: {
-          solana: {
-            createOnLogin: "users-without-wallets",
-          },
-        },
-      }}
-    >
-      <SentryUserContext />
-      <PrivyLoginBridge>{children}</PrivyLoginBridge>
-    </PrivyProvider>
-  );
-};
-
-/**
- * Bridge component that exposes Privy's login function via context.
- * Must be rendered inside PrivyProvider.
- */
-const PrivyLoginBridge: FC<{ children: ReactNode }> = ({ children }) => {
-  const { login } = usePrivy();
-  return (
-    <PrivyLoginContext.Provider value={login}>
-      {children}
-    </PrivyLoginContext.Provider>
   );
 };
